python-mongodb-rag-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
motor
openai>=1.0
pymupdf
numpy
